Rename serialized-analysis to elements-metadata
